FIX: User restrictions & toastNotification globally

+++ scripts/dashboard.js
@@ -3,24 +3,23 @@ import { getContacts } from "../api/contactAPI.js";
 import { getSuscriptions } from "../api/suscriptionAPI.js";
 
 document.addEventListener("DOMContentLoaded", async () => {
-  
-  // Obtener el nombre del usuario desde sessionStorage
-  const userName = sessionStorage.getItem('name') || 'Admin';
+  const userName = sessionStorage.getItem('name');
+
+  if (!userName) {
+    window.location.href = "./login.html";
+    return;
+  }
+
   const welcomeMessage = document.getElementById("welcomeMessage");
   if (welcomeMessage) {
     welcomeMessage.textContent = `Welcome back, ${userName}`;
   }
   
-  // Event listener para el botón de logout
   const logoutBtn = document.getElementById("logoutBtn");
   if (logoutBtn) {
     logoutBtn.addEventListener("click", (e) => {
       e.preventDefault();
-      
-      // Limpiar session storage
       sessionStorage.clear();
-      
-      // Redirigir al login
       window.location.href = "./login.html";
     });
   }
